07.10.24

+++ good-old-days/Leleki.cpp
@@ -4,9 +4,10 @@ using namespace std;
 
 int main()
 {
-    int a, b, minimum, maximum;
+    long long a, b, minimum, maximum;
     cin>>a>>b;
 
+    
     if (a == 0 || b == 0){
         cout<<"NO";
     }
